requirements.txt
numpy
cupy-cuda12x
numba